Remove problematic ensure block from Middleware
When dealing with a non-finalized middleware response, the ensure
block would execute before the middleware was finalized. The ensure
block called trace.release but the trace wasn't actually submitted
until the response was finalized. In the case of using the Middleware
probe, it would also mean that trace.done for the spans was called
after the release. This would cause these to be ignored as well as
causing the trace to be discarded as invalid, leading to a 0ms in
the reports!