Fix `mongoid` probe not activating correctly
This fixes an issue where adding `mongoid` to the probes list does
not correctly activate the relevant probe correctly. In other words
the `mongoid` probe didn't work.

Historically, older versions of the `mongoid` gem uses the `moped`
driver, and newer versions switched to using the official driver,
and we used to support both. The purpose of the `mongoid` probe was
to detect which of the two is being used, and activate either one
of the probes for the underlying driver. In either case, the actual
events are captured by the probes instrumenting the driver.

As of agent v5, we dropped support for `moped` and the versions of
`mongoid` that depends on it. Therefore, as of now, the `mongoid`
probe is just an alias for the `mongo` probe, adding either or both
to the probes list is expected to have the same result.

This, however, did not work as intended. The implementation prior
to this commit attempts to use the `Skylight::Probes.probe` API
to activate the `:mongo` probe programatically from the `install`
hook. However, this is called from `Skylight::Probes.install!`,
which pre-compute a list of "pending" probes that needs installing,
then call their `install` hook in a loop. Thus, modifying the list
inside the loop has no effect (and modifying the thing you are in
the middle of iterating over is never a good idea anyway).

In the current code, all `Skylight::Probes.probe` does (which is
what adding something to `config.skylight.probes` ultimately does)
is requiring the probe file, so we can accomplish the intended
behavior by requiring `probes/mongo.rb` from `probes/mogoid.rb`. If
the file was already required before, that will be no-op per usual
Ruby semantics.