Fix `mongo` probe not instrumenting existing clients
This fixes an issue where `mongo` events were not collected in some
cases. Specifically – if mongo clients/connections were established
prior to the Skylight initializer runs, then those clients are not
instumented by Skylight.

We uses mongo's "global" instrumentation API to collect events. In
the documentation, they mentioned that this must be done prior to
the creation of the clients/connections or else it has no effect.

To fix this, we have to loop through any existing clients in the
`install` hook and subscribe to those clients manually, in addition
to subscribing "globally".