Fix `Delayed::Job` probe crash on deserialization errors
Previously, the `Delayed::Job` probe eagerly deserializes the
`payload_object` in order to name the trace. This moves the work
ahead of where it would normally happen, and in case of an error
during deserialization (e.g. syntax error in the YAML payload, or
a deleted database record as in the customer issue), the error will
be raised outside of the block where it's normally be caught.

This fix defers the naming until a spot where we can guarantee DJ's
successful deserialization, preventing the crashes while capturing
the deserialization (database load) time within the trace.

If an error occurs prior to reaching that step, the trace will have
a default "unknown" label.

It is possible that this would affect some traces that we previously
were able to name successfully – specifically, if an error occurs in
a lifecycle hook. If this turns out to be an issue, we can try and
opportunistically check for `@payload_object` on the `job` instance
during the `:error` and `:failure` hook, but this is reaching too
deep into the implementation internals for my taste. And honestly it
seems fair enough to categorize anything that never made it to the
actual `#perform` as generic "unknown+error" traces.

Fixes #491